feat: add vaccine data

+++ src/CustomFunction/custom.js
@@ -1,8 +1,7 @@
 module.exports = {
   getNumber: function (string) {
     if (typeof string === 'string') {
-      return parseInt(string.split('').filter((el) => el !== ',').join(''));
+      return parseFloat(string.split('').filter((el) => el !== ',').join(''));
     }
   }
-  
 }
